Fix failing tests to not assume the unnamed thread ID
The tests were failing becaue they incorrectly assumed the value of the
unnamed thread ID; the thread ID turned out to be `ThreadId(5)`, not
`ThreadId(4)`. Return the unnamed thread ID from `run_log` instead, and
check for that.

In the process, simplify the log handling to retrieve the full list of
log files every time, and use `assert_eq!` to compare that to the
expected files. That makes it easier to analyze test failures, because
the `assert_eq!` will show the expected files and the actual files. This
also makes it easier to check every time that log files that shouldn't
exist don't exist.

Factor out the log prefix as a constant.