Allow for recursive log invocations
viz. when a log argument's evaluation will cause other calls to the log
macros.

This requires not taking the per-thread ref-cell while evaluating the
arguments. For the default formatter, it's rather straightforward. For
custom formatting functions, a new primitive is created that allows to
lazily get the writer, instead of borrowing it immediately and
forwarding it to the formatting function.

This is a breaking API change, so this will require a new version bump.

Signed-off-by: Benjamin Bouvier <public@benj.me>